feat: creating entities

+++ quickOS.Application/DTOs/InputModels/ServiceProvidedInputModel.cs
@@ -19,7 +19,7 @@ public class ServiceProvidedInputModel
 
     [Required(ErrorMessage = "O campo {0} é obrigatório")]
     [Range(0, 99999999.99, ErrorMessage = "O campo {0} deve ser um valor entre {1} e {2}")]
-    public double Price { get; set; }
+    public decimal Price { get; set; }
 
     [Required(ErrorMessage = "O campo {0} é obrigatório")]
     public bool IsActive { get; set; }
@@ -29,7 +29,7 @@ public class ServiceProvidedQueryParams : QueryParams
 {
     public int? Code { get; set; }
     public string? Name { get; set; } = string.Empty;
-    public double? Price { get; set; }
+    public decimal? Price { get; set; }
     public bool? IsActive { get; set; }
 }
